services/recurring-tasks-service/pyproject.toml
[project]
name = "recurring-tasks-service"
version = "2.0.0"
description = "Phase V Recurring Tasks Service - Generates task instances based on schedules"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "dapr-ext-grpc>=1.12.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.9",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.2",
    "apscheduler>=3.10.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
